docker-compose.yml: add extra host mapping for RBAC/NS images 2.0.1-1
Now, the RBAC/Cable images have an extra naming:
  - RBAC management studio: sirius-rbac.lnls.br
  - RBAC auth services: sirius-rbac-auth.lnls.br
  - Naming service: sirius-ns.lnls.br

If running on the same host, with the same
docker network, the mapping must be the internal
docker IP, such as 172.18.0.3. This is necessary
as the port mapping between the container/host
might change and the docker DNS services knows
how to deal with this. If running on the same,
but the actual IP is used, the docker DNS service
will not be used and the services will try to
use the actual specified port, which could be
incorrect due to container/host port mapping.